+++ operator/internal/manifests/memberlist.go
@@ -3,10 +3,17 @@ package manifests
 import (
 	"fmt"
 
-	"k8s.io/apimachinery/pkg/util/intstr"
+	"github.com/imdario/mergo"
 
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
+
+	lokiv1 "github.com/grafana/loki/operator/apis/loki/v1"
+)
+
+const (
+	podIPKey = "POD_IP"
 )
 
 // BuildLokiGossipRingService creates a k8s service for the gossip/memberlist members of the cluster
@@ -33,4 +40,47 @@ func BuildLokiGossipRingService(stackName string) *corev1.Service {
 			Selector: commonLabels(stackName),
 		},
 	}
+}
+
+func configureHashRingEnv(p *corev1.PodSpec, opts Options) error {
+	resetProxyVar(p, podIPKey)
+
+	if opts.Stack.HashRing == nil {
+		return nil
+	}
+
+	if opts.Stack.HashRing.Type != lokiv1.HashRingMemberList {
+		return nil
+	}
+
+	if opts.Stack.HashRing.MemberList == nil {
+		return nil
+	}
+
+	if opts.Stack.HashRing.MemberList.BindNetworkType == lokiv1.MemberListPrivateNetwork {
+		return nil
+	}
+
+	src := corev1.Container{
+		Env: []corev1.EnvVar{
+			{
+				Name: gossipBindAddrEnvVarName,
+				ValueFrom: &corev1.EnvVarSource{
+					FieldRef: &corev1.ObjectFieldSelector{
+						APIVersion: "v1",
+						FieldPath:  "status.podIP",
+					},
+				},
+			},
+		},
+	}
+
+	for i, dst := range p.Containers {
+		if err := mergo.Merge(&dst, src, mergo.WithAppendSlice); err != nil {
+			return err
+		}
+		p.Containers[i] = dst
+	}
+
+	return nil
 }